Dependecy scheduling tests.
