+++ build-coordinator/src/main/java/org/jboss/pnc/coordinator/builder/DefaultBuildCoordinator.java
@@ -48,6 +48,7 @@ import org.jboss.pnc.spi.coordinator.BuildCoordinator;
 import org.jboss.pnc.spi.coordinator.BuildScheduler;
 import org.jboss.pnc.spi.coordinator.BuildSetTask;
 import org.jboss.pnc.spi.coordinator.BuildTask;
+import org.jboss.pnc.spi.coordinator.BuildTaskRef;
 import org.jboss.pnc.spi.coordinator.CompletionStatus;
 import org.jboss.pnc.spi.coordinator.InMemory;
 import org.jboss.pnc.spi.coordinator.ProcessException;
@@ -468,7 +469,6 @@ public class DefaultBuildCoordinator implements BuildCoordinator {
         boolean temporaryBuild = buildTask.getBuildOptions().isTemporaryBuild();
         return new BuildTaskContext(
                 buildTask.getContentId(),
-                buildTask.getUser().getId().toString(),
                 temporaryBuild,
                 ExpiresDate.getTemporaryBuildExpireDate(systemConfig.getTemporaryBuildsLifeSpan(), temporaryBuild));
     }
@@ -941,6 +941,11 @@ public class DefaultBuildCoordinator implements BuildCoordinator {
                                 && Integer.valueOf(buildConfigSetRecordId).equals(t.getBuildConfigSetRecordId()))
                 .sorted(Comparator.comparing(bt -> bt.getBuildConfigurationAudited().getName()))
                 .collect(Collectors.toList());
+    }
+
+    @Override
+    public List<BuildTaskRef> getSubmittedBuildTaskRefsBySetId(int buildConfigSetRecordId) {
+        throw new UnsupportedOperationException("To be used only with the remote build coordinator.");
     }
 
     @PostConstruct
