+++ build-coordinator/src/test/java/org/jboss/pnc/coordinator/test/BuildCoordinatorFactory.java
@@ -112,7 +112,8 @@ public class BuildCoordinatorFactory {
                 null,
                 null,
                 null,
-                "true");
+                "true",
+                "3");
     }
 
     @Alternative
