+++ messaging/src/test/java/org/jboss/pnc/messaging/SysConfigProducer.java
@@ -59,6 +59,7 @@ public class SysConfigProducer {
                 null,
                 null,
                 null,
-                "true"); // TODO change to false
+                "false",
+                "3");
     }
 }
