+++ pnc-mock/src/main/java/org/jboss/pnc/mock/common/SystemConfigMock.java
@@ -59,6 +59,7 @@ public class SystemConfigMock {
                 null,
                 null,
                 null,
-                "true"); // TODO change to false
+                "false",
+                "3");
     }
 }
