+++ remote-build-coordinator/pom.xml
@@ -58,12 +58,10 @@
     <dependency>
       <groupId>org.jboss.pnc.rex</groupId>
       <artifactId>dto</artifactId>
-      <version>0.1-SNAPSHOT</version>
     </dependency>
     <dependency>
       <groupId>org.jboss.pnc.rex</groupId>
-      <artifactId>core</artifactId>
-      <version>0.1-SNAPSHOT</version>
+      <artifactId>rex-api</artifactId>
     </dependency>
     <dependency>
       <groupId>org.jboss.pnc</groupId>
