+++ remote-build-coordinator/src/main/java/org/jboss/pnc/remotecoordinator/builder/datastore/DatastoreAdapter.java
@@ -564,7 +564,7 @@ public class DatastoreAdapter {
                 alignmentPreference,
                 processedDependenciesCache,
                 noRebuildCause);
-        return Optional.of(wrapper.get());
+        return Optional.ofNullable(wrapper.get());
     }
 
     @Deprecated
