+++ remote-build-coordinator/src/test/java/org/jboss/pnc/remotecoordinator/test/ConfigurationsTest.java
@@ -19,18 +19,7 @@ package org.jboss.pnc.remotecoordinator.test;
 
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
-import org.jboss.arquillian.junit.InSequence;
-import org.jboss.pnc.enums.BuildCoordinationStatus;
-import org.jboss.pnc.enums.BuildStatus;
-import org.jboss.pnc.enums.RebuildMode;
-import org.jboss.pnc.model.BuildConfiguration;
-import org.jboss.pnc.model.BuildConfigurationSet;
-import org.jboss.pnc.model.User;
-import org.jboss.pnc.spi.BuildOptions;
-import org.jboss.pnc.spi.coordinator.BuildTask;
-import org.jboss.pnc.spi.coordinator.BuildCoordinator;
-import org.jboss.pnc.spi.coordinator.BuildSetTask;
-import org.jboss.pnc.spi.coordinator.Remote;
+import org.jboss.pnc.mock.model.builders.TestProjectConfigurationBuilder;
 import org.jboss.shrinkwrap.api.spec.JavaArchive;
 import org.junit.Assert;
 import org.junit.Test;
@@ -38,65 +27,49 @@ import org.junit.runner.RunWith;
 
 import javax.inject.Inject;
 import javax.persistence.PersistenceException;
-import java.util.Set;
-
-import static org.assertj.core.api.Assertions.assertThat;
 
 /**
  * Created by <a href="mailto:matejonnet@gmail.com">Matej Lazar</a> on 2014-11-23.
  */
 @RunWith(Arquillian.class)
-public class ConfigurationsTest extends ProjectBuilder {
+public class ConfigurationsTest {
+    // TODO make sure there is a test for REJECTED build in the RemoteBuildCoordinatorTest
 
     @Deployment
     public static JavaArchive createDeployment() {
         return BuildCoordinatorDeployments.deployment(
-                BuildCoordinatorDeployments.Options.WITH_DATASTORE,
-                BuildCoordinatorDeployments.Options.WITH_BPM);
+                BuildCoordinatorDeployments.Options.WITH_DATASTORE);
     }
 
     @Inject
-    @Remote
-    BuildCoordinator buildCoordinator;
+    TestProjectConfigurationBuilder configurationBuilder;
 
-    @Test(expected = PersistenceException.class) // TODO test is not run as expected exception is thrown
-                                                 // configurationBuilder.build...
-    @InSequence(10)
+    @Test
     public void dependsOnItselfConfigurationTestCase() throws Exception {
-
-        BuildConfiguration buildConfiguration = configurationBuilder.buildConfigurationWhichDependsOnItself();
-
-        User user = User.Builder.newBuilder().id(1).build();
-
-        BuildOptions buildOptions = new BuildOptions();
-        buildOptions.setBuildDependencies(false);
-
-        BuildSetTask taskSet = buildCoordinator.buildConfig(buildConfiguration, user, buildOptions);
-        Set<BuildTask> buildTasks = taskSet.getBuildTasks();
-        assertThat(buildTasks).hasSize(1);
-        BuildTask buildTask = buildTasks.iterator().next();
-        Assert.assertEquals(BuildCoordinationStatus.REJECTED, buildTask.getStatus());
-        Assert.assertTrue(
-                "Invalid status description: " + buildTask.getStatusDescription(),
-                buildTask.getStatusDescription().contains("itself"));
+        try {
+            configurationBuilder.buildConfigurationWhichDependsOnItself();
+        } catch (PersistenceException e) {
+            String message = "itself";
+            Assert.assertTrue(
+                    "Expected exception message to contain " + message,
+                    e.getMessage().contains(message));
+            return;
+        }
+        Assert.fail("Did not receive expected exception.");
     }
 
-    @Test(expected = PersistenceException.class) // TODO test is not run as expected exception is thrown
-                                                 // configurationBuilder.build...
-    @InSequence(15)
+    @Test
     public void cycleConfigurationTestCase() throws Exception {
-
-        BuildConfigurationSet buildConfigurationSet = configurationBuilder.buildConfigurationSetWithCycleDependency();
-
-        User user = User.Builder.newBuilder().id(1).build();
-
-        BuildOptions buildOptions = new BuildOptions();
-        buildOptions.setRebuildMode(RebuildMode.FORCE);
-        BuildSetTask buildSetTask = buildCoordinator.buildSet(buildConfigurationSet, user, buildOptions);
-        Assert.assertEquals(BuildStatus.REJECTED, buildSetTask.getTaskStatus());
-        Assert.assertTrue(
-                "Invalid status description: " + buildSetTask.getStatusDescription(),
-                buildSetTask.getStatusDescription().contains("Cycle dependencies found"));
+        try {
+            configurationBuilder.buildConfigurationSetWithCycleDependency();
+        } catch (PersistenceException e) {
+            String message = "circular reference";
+            Assert.assertTrue(
+                    "Expected exception message to contain [" + message + "] but it has [" + e.getMessage() + "]",
+                    e.getMessage().contains(message));
+            return;
+        }
+        Assert.fail("Did not receive expected exception.");
     }
 
 }
