+++ remote-build-coordinator/src/test/java/org/jboss/pnc/remotecoordinator/test/WaitForDependencyBuildTest.java
@@ -17,79 +17,63 @@
  */
 package org.jboss.pnc.remotecoordinator.test;
 
-import org.jboss.pnc.common.json.ConfigurationParseException;
-import org.jboss.pnc.model.BuildConfiguration;
-import org.jboss.pnc.enums.BuildCoordinationStatus;
-import org.jboss.pnc.spi.coordinator.BuildTask;
-import org.jboss.pnc.spi.datastore.DatastoreException;
-import org.jboss.pnc.spi.exception.CoreException;
-import org.jboss.pnc.test.util.Wait;
-import org.junit.Before;
 import org.junit.Ignore;
-import org.junit.Test;
 
-import java.time.temporal.ChronoUnit;
-import java.util.List;
-import java.util.Optional;
-import java.util.concurrent.TimeoutException;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-@Ignore // SHOULD BE DONE IN INTEGRATION TESTS WITH REX
-public class WaitForDependencyBuildTest extends AbstractDependentBuildTest {
-
-    private BuildConfiguration configParent;
-    private BuildConfiguration configDependency;
-
-    private MockBuildSchedulerWithManualBuildCompletion buildScheduler = new MockBuildSchedulerWithManualBuildCompletion();
-
-    @Before
-    public void initialize() throws DatastoreException, ConfigurationParseException {
-        configDependency = buildConfig("Dependency");
-        configParent = buildConfig("Parent", configDependency);
-
-        super.initialize();
-
-        saveConfig(configDependency);
-        saveConfig(configParent);
-    }
-
-    @Test
-    public void shouldNotStartParentBuildWhenDependencyIsRunning()
-            throws CoreException, TimeoutException, InterruptedException {
-
-        // start dependency
-        build(configDependency);
-        Wait.forCondition(() -> buildScheduler.isBuilding(configDependency.getId()), 3, ChronoUnit.SECONDS);
-
-        // start parent while dependency is running
-        build(configParent);
-
-        // parent should wait for dependency to complete
-        BuildTask parentBuildTask = getSubmittedBuildTaskByConfigurationId(configParent.getId()).get();
-        assertThat(parentBuildTask.getStatus()).isEqualTo(BuildCoordinationStatus.WAITING_FOR_DEPENDENCIES);
-
-        // complete the dependency
-        BuildTask dependencyBuildTask = getScheduledBuildTaskByConfigurationId(configDependency.getId()).get();
-        buildScheduler.completeBuild(dependencyBuildTask.getId());
-
-        // check if parent has started
-        Wait.forCondition(
-                () -> buildScheduler.isBuilding(parentBuildTask.getBuildConfigurationAudited().getId()),
-                3,
-                ChronoUnit.SECONDS);
-
-        buildScheduler.completeBuild(parentBuildTask.getId());
-
-        List<BuildConfiguration> configsWithTasks = getBuiltConfigs();
-        assertThat(configsWithTasks).isEmpty();
-    }
-
-    private Optional<BuildTask> getSubmittedBuildTaskByConfigurationId(Integer buildConfigurationId) {
-        return coordinator.getSubmittedBuildTasks()
-                .stream()
-                .filter(bt -> bt.getBuildConfigurationAudited().getId().equals(buildConfigurationId))
-                .findAny();
-    }
+@Ignore //TODO SHOULD BE DONE IN INTEGRATION TESTS WITH REX
+public class WaitForDependencyBuildTest  {
+//extends AbstractDependentBuildTest
+//    private BuildConfiguration configParent;
+//    private BuildConfiguration configDependency;
+//
+//    private MockBuildSchedulerWithManualBuildCompletion buildScheduler = new MockBuildSchedulerWithManualBuildCompletion();
+//
+//    @Before
+//    public void initialize() throws DatastoreException, ConfigurationParseException {
+//        configDependency = buildConfig("Dependency");
+//        configParent = buildConfig("Parent", configDependency);
+//
+//        super.initialize();
+//
+//        saveConfig(configDependency);
+//        saveConfig(configParent);
+//    }
+//
+//    @Test
+//    public void shouldNotStartParentBuildWhenDependencyIsRunning()
+//            throws CoreException, TimeoutException, InterruptedException {
+//
+//        // start dependency
+//        build(configDependency);
+//        Wait.forCondition(() -> buildScheduler.isBuilding(configDependency.getId()), 3, ChronoUnit.SECONDS);
+//
+//        // start parent while dependency is running
+//        build(configParent);
+//
+//        // parent should wait for dependency to complete
+//        BuildTask parentBuildTask = getSubmittedBuildTaskByConfigurationId(configParent.getId()).get();
+//        assertThat(parentBuildTask.getStatus()).isEqualTo(BuildCoordinationStatus.WAITING_FOR_DEPENDENCIES);
+//
+//        // complete the dependency
+//        BuildTask dependencyBuildTask = getScheduledBuildTaskByConfigurationId(configDependency.getId()).get();
+//        buildScheduler.completeBuild(dependencyBuildTask.getId());
+//
+//        // check if parent has started
+//        Wait.forCondition(
+//                () -> buildScheduler.isBuilding(parentBuildTask.getBuildConfigurationAudited().getId()),
+//                3,
+//                ChronoUnit.SECONDS);
+//
+//        buildScheduler.completeBuild(parentBuildTask.getId());
+//
+//        List<BuildConfiguration> configsWithTasks = getBuiltConfigs();
+//        assertThat(configsWithTasks).isEmpty();
+//    }
+//
+//    private Optional<BuildTask> getSubmittedBuildTaskByConfigurationId(Integer buildConfigurationId) {
+//        return coordinator.getSubmittedBuildTasks()
+//                .stream()
+//                .filter(bt -> bt.getBuildConfigurationAudited().getId().equals(buildConfigurationId))
+//                .findAny();
+//    }
 
 }
