+++ spi/src/main/java/org/jboss/pnc/spi/coordinator/BuildCoordinator.java
@@ -44,7 +44,7 @@ public interface BuildCoordinator {
             User user,
             BuildOptions buildOptions) throws BuildConflictException, CoreException;
 
-    @Deprecated //used only in the tests
+    @Deprecated // used only in the tests
     BuildSetTask buildSet(BuildConfigurationSet buildConfigurationSet, User user, BuildOptions buildOptions)
             throws CoreException;
 
@@ -52,7 +52,7 @@ public interface BuildCoordinator {
             BuildConfigurationSet buildConfigurationSet,
             Map<Integer, BuildConfigurationAudited> buildConfigurationAuditedsMap,
             User user,
-            BuildOptions buildOptions) throws CoreException, BuildConflictException;
+            BuildOptions buildOptions) throws CoreException;
 
     Optional<BuildTask> getSubmittedBuildTask(String buildId);
 
@@ -81,7 +81,7 @@ public interface BuildCoordinator {
 
     boolean cancelSet(int buildConfigSetRecordId) throws CoreException;
 
-    @Deprecated //used only internally
+    @Deprecated // used only internally
     void updateBuildTaskStatus(BuildTask task, BuildCoordinationStatus status);
 
     void updateBuildConfigSetRecordStatus(BuildConfigSetRecord setRecord, BuildStatus status, String description)
